pi/requirements.txt
# pip requirements
pyserial==3.5
psycopg2-binary==2.9.10
pyyaml==6.0.2
flask==3.1.0
sqlalchemy==2.0.41
flask_migrate==4.1.0
Flask-SQLAlchemy==3.1.1
urllib3==2.4.0